Add custom HR border, region'd  sections thus far
Featuring an SVG I created using VSCode's "add file" button that it now does not detect as being untracked or changed or anything despite what the command line says! Let's see if it gets included in this commit after manually `git add`ing it, or if I have to manually `git commit` it, too. :)