Implement/style skill subtypes, minor name changes
This commit started out with making a function to clone a newly added "inner" node in each relevant skill row, which mostly worked, but didn't quite have *all* the bindings right.
From there, I discovered that v-for isn't a one and done iteration; it live updates, like everything else, when new things are added to the array it pulls from. So, at present, adding new subtypes is as simple as adding them to the skill data (given the new HTML structure).