Make skill headers not <h>s, use aria-labelledby instead
It makes some sense to have the skill "headers" be <h>s, but <h>s are kind of more for navigation/calling out what a block of content immediately following it is, and these don't do that.
Meanwhile, using aria-labelledby makes the connection between these "headers" and their column-mates more than just visual.