Refactor setTheme and getScoreX functions
While working on adding the feats section, I learned how to pass references to elements to JS from HTML via Vue, which got me looking at the setTheme function. I quickly found myself picking it apart and reorganizing it.
The score function refactor was before and mostly unrelated to this little escapade, but it's fitting to include regardless, since it's small and doesn't break anything.